chore: Merge fixtures in a single file
